+++ examples/Cargo.toml
@@ -13,8 +13,8 @@ edition = "2018"
 default = ["native-tls", "schema", "kubederive", "ws"]
 kubederive = ["kube/derive"] # by default import kube-derive with its default features
 schema = ["kube-derive/schema"] # crd_derive_no_schema shows how to opt out
-native-tls = ["kube/native-tls", "kube-runtime/native-tls"]
-rustls-tls = ["kube/rustls-tls", "kube-runtime/rustls-tls"]
+native-tls = ["kube/native-tls"]
+rustls-tls = ["kube/rustls-tls"]
 ws = ["kube/ws"]
 
 [dev-dependencies]
@@ -23,7 +23,7 @@ env_logger = "0.8.2"
 futures = "0.3.8"
 kube = { path = "../kube", version = "^0.55.0", default-features = false, features = ["admission"] }
 kube-derive = { path = "../kube-derive", version = "^0.55.0", default-features = false } # only needed to opt out of schema
-kube-runtime = { path = "../kube-runtime", version = "^0.55.0", default-features = false }
+kube-runtime = { path = "../kube-runtime", version = "^0.55.0" }
 kube-core = { path = "../kube-core", version = "^0.55.0", default-features = false }
 k8s-openapi = { version = "0.11.0", features = ["v1_20"], default-features = false }
 log = "0.4.11"
