+++ kube-runtime/Cargo.toml
@@ -14,7 +14,7 @@ edition = "2018"
 
 [dependencies]
 futures = "0.3.8"
-kube = { path = "../kube", version = "^0.55.0", default-features = false }
+kube = { path = "../kube", version = "^0.55.0", default-features = false, features = ["client"] }
 derivative = "2.1.1"
 serde = "1.0.118"
 smallvec = "1.6.0"
@@ -27,11 +27,6 @@ tokio-util = { version = "0.6.0", features = ["time"] }
 [dependencies.k8s-openapi]
 version = "0.11.0"
 default-features = false
-
-[features]
-default = ["native-tls"]
-native-tls = ["kube/native-tls"]
-rustls-tls = ["kube/rustls-tls"]
 
 [dev-dependencies]
 kube-derive = { path = "../kube-derive", version = "^0.55.0"}
