+++ kube/Cargo.toml
@@ -54,7 +54,6 @@ tokio-native-tls = { version = "0.3.0", optional = true }
 tokio-rustls = { version = "0.22.0", features = ["dangerous_configuration"], optional = true }
 bytes = { version = "1.0.0", optional = true }
 tokio = { version = "1.0.1", features = ["time", "signal", "sync"], optional = true }
-static_assertions = "1.1.0"
 kube-derive = { path = "../kube-derive", version = "^0.55.0", optional = true }
 kube-core = { path = "../kube-core", version = "^0.55.0"}
 jsonpath_lib = { version = "0.3.0", optional = true }
