+++ kube/src/service/mod.rs
@@ -4,7 +4,6 @@ mod auth;
 #[cfg(feature = "gzip")] mod compression;
 mod headers;
 mod log;
-mod tls;
 mod url;
 
 #[cfg(feature = "gzip")]
@@ -13,6 +12,5 @@ pub(crate) use self::{
     auth::{AuthLayer, Authentication},
     headers::set_default_headers,
     log::LogRequest,
-    tls::HttpsConnector,
     url::set_cluster_url,
 };
